+++ pkg/task/table.go
@@ -1,8 +1,6 @@
 package task
 
 import (
-	"fmt"
-
 	"github.com/tooploox/oya/pkg/types"
 )
 
@@ -29,17 +27,40 @@ func (tt Table) AddTask(name Name, task Task) {
 
 func (tt Table) AddDoc(taskName Name, s string) {
 	tt.meta[taskName] = Meta{
-		Doc: s,
+		Doc:              s,
+		OriginalTaskName: taskName,
 	}
 }
 
 func (tt Table) ImportTasks(alias types.Alias, other Table) {
-	for key, t := range other.tasks {
+	for name, t := range other.tasks {
 		// TODO: Detect if task already set.
-		tt.AddTask(Name(fmt.Sprintf("%v.%v", alias, key)), t)
+		tt.AddTask(Name(name).Aliased(alias), t)
 	}
 }
 
+// Expose copies tasks under an alias to global scope (without the alias)
+// never overriding the existing global tasks.
+func (tt Table) Expose(alias types.Alias) {
+	for name, task := range tt.tasks {
+		if name.IsAliased(alias) {
+			globalName := name.Unaliased()
+			_, ok := tt.LookupTask(globalName)
+			if !ok {
+				meta := tt.meta[name]
+				meta.OriginalTaskName = name
+				tt.addTaskWithMeta(globalName, task, meta)
+			}
+		}
+	}
+}
+
+func (tt Table) addTaskWithMeta(name Name, task Task, meta Meta) {
+	tt.tasks[name] = task
+	tt.meta[name] = meta
+}
+
+// ForEach invokes the callback for each task.
 func (tt Table) ForEach(f func(taskName Name, task Task, meta Meta) error) error {
 	for taskName, task := range tt.tasks {
 		meta := tt.meta[taskName]
@@ -50,13 +71,14 @@ func (tt Table) ForEach(f func(taskName Name, task Task, meta Meta) error) error
 	return nil
 }
 
+// ForEachSorted invokes the callback for each task, the tasks sorted alphabetically.
 func (tt Table) ForEachSorted(f func(taskName Name, task Task, meta Meta) error) error {
 	taskNames := make([]Name, 0, len(tt.tasks))
 	for taskName := range tt.tasks {
 		taskNames = append(taskNames, taskName)
 	}
-
 	Sort(taskNames)
+
 	for _, taskName := range taskNames {
 		task := tt.tasks[taskName]
 		meta := tt.meta[taskName]